uint64_t src organizing part2

+++ src/hash.h
@@ -5,12 +5,12 @@
 
 typedef struct tuple{
   int32_t key;
-  int32_t payload;
+  uint64_t payload;
 } tuple;
 
 typedef struct relation{
   tuple *tuples;
-  uint32_t num_tuples;
+  uint64_t num_tuples;
 } relation;
